+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/listbox/DefaultListBoxBuilder.java
@@ -40,7 +40,6 @@ import java.util.function.Function;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-import static is.codion.swing.common.model.component.combobox.FilterComboBoxModel.filterComboBoxModel;
 import static java.util.Collections.emptySet;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toCollection;
@@ -67,7 +66,7 @@ final class DefaultListBoxBuilder<T>
 
 	@Override
 	protected JComboBox<T> createComponent() {
-		FilterComboBoxModel<T> comboBoxModel = filterComboBoxModel();
+		FilterComboBoxModel<T> comboBoxModel = FilterComboBoxModel.<T>builder().build();
 		ListComboBox<T> comboBox = new ListComboBox<>(comboBoxModel, itemValue, linkedValue);
 		comboBox.setEditor(new Editor<>(itemValue));
 		comboBox.setEditable(true);